requirements.txt
nose
sphinx
pexpect
paho-mqtt
pypubsub
